requirements.txt
fpdf2>=2.7.8
qrcode[pil]>=7.4.2